requirements.txt
asgiref==3.5.2
Django==4.1.3
django-tinymce==3.5.0
python-dotenv==0.21.0
sqlparse==0.4.3
